feat(skill-doctor): warn when a skill declares no metadata.triggers
The trigger parser moves to src/hooks/lib/skill-triggers.ts so the index
generator and the doctor share one reader (src/tools may not import from
src/targets). Zero triggers warns; fewer than three warns as thin; a
top-level triggers: key still counts as none, since that key fails skill
packaging.

Co-authored by Jarvis · https://github.com/kovrichard/portable-agent-layer
